拆分工具类
